UI Updates Going On
Started getting rid of metroframework
