+++ Project_OpenSource/CoreCheat-CSGO-Reborn-master/Program.cs
@@ -11,7 +11,8 @@ namespace CoreCheat_Reborn
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new Version_Selector.VersionSelector());
+            //Application.Run(new Version_Selector.VersionSelector());
+            Application.Run(new Versions.Panorama.Main());
         }
     }
 }
